Completed direct AN bindings
Finished what appears to be a working version of direct bindings to the
functions defined in model_Synapse.c that handle firing rate
calculations (Syanpse and SingleAN). No spiking is simulated yet, since I
haven't figured out how to provide the correct random numbers yet. Fractional
Gaussian noise is also not implemented, since I still need to write the Julia
function for fractional Gaussian noise.

The main additions are direct bindings to Synapse and SingleAN, but these are
not intended for users. Next step will be to write a set of wrappers in Julia
that handle all of the ugly bits for the user!