Added tests to check frequency tuning
Noticed a biazrre issue wherein AN/synapse resposnes are completely
wrong as a function of CF. Frequency tuning is all out of whack. It's
probably related to the decimate function, if I had to take a guess ---
need to investigate further. Added tests to check that frequency tuning
is reasonable and IHC passes just fine, so once AN passes here I'll know
I've fixed the issue.