Mostly fixed issue
The issues I was having with ANF/Synapse were mostly related to me
calling these functions directly on sound pressure waveforms instead
of IHC outputs. Big user error!

However, there is a new problem that has cropped up related to
spontaneous rates sometimes generating NaNs. Irritatingly and
bizarrely, it is non-deterministic and appeared to depend on the caller.
When running from REPL, it never shows up, but when running the tests
automatically, it shows up frequently. Not sure what gives.