Updated license and switched to libzbc2014 dependency
Updated license to MIT (which appears to be more of the norm for Julia
packages). Also removed dependency on hard-coded location for Zilany,
Bruce, and Carney (2014) model and instead added a dependency on
the compiled libzbc2014 jll package. Eventually, this package will be
hosted online alongside AuditoryNerveFiber.jl and should make
installation completely seamless.