Added a workable upsampling function based on DSP.resample
Added a mini wrapper around DSP.resample for upsampling that takes
the output from DSP.resample and simply pads it with a few zeros to get
it to match in length with the expected output. This is a bit of a
bandaid fix but it's preferable to rolling a custom upsampling algorithm
(which I don't really want to optimize).