Revised several function signatures to support AbstractArray
Generalized several function signatures to dispatch on AbstractArray
or AbstractVector, instead of specific Array... Also added
several methods for each of sim_ihc_zbc2014, sim_an_zbc2014, and
sim_synapse_zbc2014 that allow dispatch over n-dimensional arrays of
inputs. In the long-term a lot of this sort of stuff should probably
live in a separate package but for the time being they are nice utility
methods to have.