Generalized most low-level functions to vector inputs
Wrote a set of macros to generate methods for all functions of the form
func(input, cf; kwargs...) such that input and cf can be either single
elements or vectors of elements. In the case of a vector of inputs,
a vector of outputs is returned. In the case of a vector of cfs, a matrix
output is returned (where each cf is a different row).

Some additional docstrings will need to be written and the docs will
need to be updated to reflect all these new features!