Added an additional macro to handle matrix -> matrix case
Added an additional dispatch paradigm to functions with signatures like
func(input, cf; kwargs...). When a vector of CFs is passed with a single
input, then a matrix (neurogram) is returned. When this matrix and the
same vector of cfs are in turn passed to a similar function, they are now
correctly handled, so one can arbitrarily compose series of such neurograms.
Additionally, one can now also pass vectors of such matrices and each
is processed in turn. Additional tests were added to handle these cases.