Extended dispatch to include matrix inputs
All simulation functions of canonical form: func(input, cf; kwargs...)
now have been extended to dispatch over matrix-type inputs or vectors
of such inputs. In these cases, we assume that the inputs represent
neurograms and thus cf should be a vector encoding the cf of each row (?)
of the matrix. Additional documentation will be needed to explain the
different possibilities, but they seem to work quite well and are already
tested in the test suite.