Switched from DSP.jl polyphase downsampling to FFT downsampling
Previoulsy, I used DSP.resample() to perform resampling. This algorithm,
from what I can tell, implements some sort of polyphase FIR downsampling.
Although a good idea in theory, the implementation was abysmally slow
due to very slow calls to kaiser() which in turn made very slow calls to
a function for calculating values of a Bessel function. This was replaced
with an algorithm inspired by scipy.signal.downsample, which simply
takes the FFT of the signal, truncates above a cutoff frequency, and
calculates the IFFT of the truncated spectrum. This seems to perform
okay (in terms of accuracy) although I haven't tested it extensively.
Importantly, it's *significantly* faster. Shortly I'll try to develop
some proper benchamarks on a decent PC to compare the implementation
here to the implementation in cochlea/apcmodels.